Add update for multiple labels + fix tests
