+++ e2e/llm_util_test/test_schema_default_with_relationship_props/test.yml
@@ -1,5 +1,4 @@
 query: >
     CALL llm_util.schema() YIELD schema RETURN schema;
 output:
-  - schema: "Node properties are the following:\nNode name: 'Person', Node properties: [{'property': 'name', 'type': 'str'}, {'property': 'age', 'type': 'int'}]\nNode name: 'Person:Student', Node properties: [{'property': 'name', 'type': 'str'}, {'property': 'age', 'type': 'int'}, {'property': 'year', 'type': 'str'}]\nNode name: 'University', Node properties: [{'property': 'name', 'type': 'str'}]\nNode name: 'City', Node properties: [{'property': 'name', 'type': 'str'}]\n\nRelationship properties are the following:\nRelationship Name: 'IS_FRIENDS_WITH', Relationship Properties: [{'property': 'since', 'type': 'str'}]\n\nThe relationships are the following:\n['(:Person)-[:IS_FRIENDS_WITH]->(:Person:Student)']\n['(:Person)-[:LIVES_IN]->(:City)']\n['(:Person:Student)-[:STUDIES_AT]->(:University)']\n['(:Person:Student)-[:LIVES_IN]->(:City)']\n"
- 
+  - schema: "Node properties are the following:\nNode name: 'Person', Node properties: [{'property': 'name', 'type': 'str'}, {'property': 'age', 'type': 'int'}, {'property': 'year', 'type': 'str'}]\nNode name: 'Student', Node properties: [{'property': 'name', 'type': 'str'}, {'property': 'age', 'type': 'int'}, {'property': 'year', 'type': 'str'}]\nNode name: 'University', Node properties: [{'property': 'name', 'type': 'str'}]\nNode name: 'City', Node properties: [{'property': 'name', 'type': 'str'}]\n\nRelationship properties are the following:\nRelationship Name: 'IS_FRIENDS_WITH', Relationship Properties: [{'property': 'since', 'type': 'str'}]\n\nThe relationships are the following:\n['(:Person)-[:IS_FRIENDS_WITH]->(:Person)']\n['(:Person)-[:IS_FRIENDS_WITH]->(:Student)']\n['(:Person)-[:LIVES_IN]->(:City)']\n['(:Person)-[:STUDIES_AT]->(:University)']\n['(:Student)-[:STUDIES_AT]->(:University)']\n['(:Student)-[:LIVES_IN]->(:City)']\n"
