+++ e2e/llm_util_test/test_schema_raw_ignore_case/test.yml
@@ -1,5 +1,5 @@
 query: >
-    CALL llm_util.schema('RaW') YIELD schema WITH schema.node_props AS nodes WITH nodes['Person:Student'] AS properties UNWIND properties AS property RETURN property["property"] AS prop, property["type"] AS type ORDER BY prop ASC;
+    CALL llm_util.schema('RaW') YIELD schema WITH schema.node_props AS nodes WITH nodes['Student'] AS properties UNWIND properties AS property RETURN property["property"] AS prop, property["type"] AS type ORDER BY prop ASC;
 output:
   - prop: "age"
     type: "int"
