+++ e2e/llm_util_test/test_schema_raw_node_properties_2/test.yml
@@ -5,3 +5,5 @@ output:
     type: "int"
   - prop: "name"
     type: "str"
+  - prop: "year"
+    type: "str"
