+++ e2e/llm_util_test/test_schema_raw_relationships/test.yml
@@ -3,13 +3,19 @@ query: >
 output:
   - start: "Person"
     type: "IS_FRIENDS_WITH"
-    end: "Person:Student"
+    end: "Person"
+  - start: "Person"
+    type: "IS_FRIENDS_WITH"
+    end: "Student"
   - start: "Person"
     type: "LIVES_IN"
     end: "City"
-  - start: "Person:Student"
+  - start: "Person"
+    type: "STUDIES_AT"
+    end: "University"
+  - start: "Student"
     type: "LIVES_IN"
     end: "City"
-  - start: "Person:Student"
+  - start: "Student"
     type: "STUDIES_AT"
     end: "University"
