+++ python/mage/llm_util/__init__.py
@@ -1,0 +1,1 @@
+from mage.llm_util.parameters import Parameter, OutputType  # noqa: F401, F402, F403
