+++ python/mage/llm_util/parameters.py
@@ -1,0 +1,16 @@
+from enum import Enum
+
+
+class Parameter(Enum):
+    END = "end"
+    NODE_PROPS = "node_props"
+    PROPERTY = "property"
+    RELATIONSHIPS = "relationships"
+    REL_PROPS = "rel_props"
+    START = "start"
+    TYPE = "type"
+
+
+class OutputType(Enum):
+    RAW = "raw"
+    PROMPT_READY = "prompt_ready"
